requirements.txt
Django==5.1.7
python-environ==0.4.54
yookassa==3.5.0
pillow==11.1.0
django-phonenumber-field==8.0.0
phonenumbers==8.13.54